cambios en el front, y agregado de seguridad y alertas

+++ src/main/java/com/backend/clinicaodontologica/service/impl/OdontologService.java
@@ -7,6 +7,7 @@ import com.backend.clinicaodontologica.dto.salida.odontologo.OdontologoSalidaDto
 
 import com.backend.clinicaodontologica.entity.Odontologo;
 
+import com.backend.clinicaodontologica.exceptions.ResourceNotFoundException;
 import com.backend.clinicaodontologica.repository.OdontologoRepository;
 import com.backend.clinicaodontologica.utils.JsonPrinter;
 import org.modelmapper.ModelMapper;
@@ -67,8 +68,14 @@ public class OdontologService implements IOdontologoService {
     }
 
     @Override
-    public void eliminarOdontologo(Long id) {
-        odontologoRepository.deleteById(id);
+    public void eliminarOdontologo(Long id)throws ResourceNotFoundException {
+        if (buscarOdontologoPorId(id) != null) {
+            odontologoRepository.deleteById(id);
+            LOGGER.warn("Se ha eliminado el odontologo con id: {}", id);
+        } else {
+            LOGGER.error("No se ha encontrado el odontologo con id {}", id);
+            throw new ResourceNotFoundException("No se ha encontrado el odontologo con id " + id);
+        }
     }
 
     public OdontologoRepository getOdontologoRepository() {
